Move all files back to same package level
