+++ clowder/breed.py
@@ -7,7 +7,8 @@ class Breed(object):
     def __init__(self, rootDirectory, url):
         self.rootDirectory = rootDirectory
         self.setupClowderDirectory(url)
-        self.configurePeru()
+        # self.configurePeru()
+        self.createSymlinks()
 
     def setupClowderDirectory(self, url):
         dotClowderDirectory = os.path.join(self.rootDirectory, '.clowder')
@@ -24,8 +25,14 @@ class Breed(object):
         clowder.utilities.ex(command)
 
     def createSymlinks(self):
-        pass
-        
+        os.chdir(self.rootDirectory)
+        clowderYAML = '.clowder/repo/clowder.yaml'
+        peruYAML = '.clowder/repo/peru.yaml'
+        if os.path.isfile(clowderYAML):
+            os.symlink(clowderYAML, 'clowder.yaml')
+        if os.path.isfile(peruYAML):
+            os.symlink(peruYAML, 'peru.yaml')
+
     def configurePeru(self):
         print('Updating peru.yaml')
         clowderDirectory = os.path.join(self.rootDirectory, '.clowder/clowder')
