+++ clowder/herd.py
