+++ clowder/remote.py
